+++ CHANGELOG.md
@@ -4,7 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+
+## [1.5.0] - 2024-04-10
+### Added
+* New language available: Arabic (MSA) (`'ar'`). Add language code constants and tests.
+
+  Note: older library versions also support the new language, this update only
+  adds new code constants.
+### Fixed
+* Change document upload to use the path `/v2/document` instead of `/v2/document/` (no trailing `/`).
+  Both paths will continue to work in the v2 version of the API, but `/v2/document` is the intended one.
+
+
+## [1.4.0] - 2023-11-03
+### Added
+* Add optional `context` parameter for text translation, that specifies
+  additional context to influence translations, that is not translated itself.
 ### Fixed
 * Remove unused `commons-math` dependency
 
@@ -89,7 +104,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial version.
 
 
-[Unreleased]: https://github.com/DeepLcom/deepl-java/compare/v1.3.0...HEAD
+[1.5.0]: https://github.com/DeepLcom/deepl-java/compare/v1.4.0...v1.5.0
+[1.4.0]: https://github.com/DeepLcom/deepl-java/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/DeepLcom/deepl-java/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/DeepLcom/deepl-java/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/DeepLcom/deepl-java/compare/v1.0.1...v1.1.0